Reapply "Correcciones y prevención de errores en endpoints"
This reverts commit 699f415465ad2ce7b6b0a99f54a7588abd99b0f5.

+++ conecta-T.Api/routes/publicacion.js
@@ -5,6 +5,10 @@ const router = Router();
 
 // Metodo GET --> /publicaciones (Obtener lista de publicaciones de quienes sigue un usuario -FEED-)
 router.get('/publicaciones', async (req, res) => {
+
+    if (!req.body) {
+        return res.status(400).json({"error": "Debe proporcionar el Id del usuario en sesion"});
+    }
     const { usuarioId } = req.body;
 
     if (typeof usuarioId != 'number') {
@@ -17,7 +21,7 @@ router.get('/publicaciones', async (req, res) => {
         });
 
         if (!usuario) {
-            res.status(404).json({"error": "No se encontró un usuario con ese Id"});
+            return res.status(404).json({"error": "No se encontró un usuario con ese Id"});
         }
         const usuarioFeed = await prisma.usuario.findUnique({
             where: { id: usuarioId },
@@ -41,6 +45,14 @@ router.get('/publicaciones/:autorId', async (req, res) => {
     const autorId = parseInt(req.params.autorId);
 
     try {
+        const usuario = await prisma.usuario.findUnique({
+            where: { id: autorId }
+        });
+
+        if (!usuario) {
+            return res.status(404).json({"error": "No se encontró un usuario con ese Id"});
+        }
+
         const publicaciones = await prisma.publicacion.findMany({
             where: { autorId },
             orderBy: { fechaCreacion: "desc" }
@@ -54,7 +66,19 @@ router.get('/publicaciones/:autorId', async (req, res) => {
 // Metodo POST --> /publicaciones (Subir nueva publicacion)
 router.post('/publicaciones', async (req, res) => {
     const { texto, foto, autorId } = req.body;
+
+    if (typeof autorId != 'number') {
+        return res.status(400).json({"error": "El Id del autor debe ser un número"});
+    }
+
     try {
+        const usuario = await prisma.usuario.findUnique({
+            where: { id: usuarioId }
+        });
+
+        if (!usuario) {
+            return res.status(404).json({"error": "No se pudo encontrar un usuario con ese Id"});
+        }
         /*
         Ejemplo: Ingresar esto en el body
         {
@@ -91,6 +115,14 @@ router.put('/publicaciones/:id', async (req, res) => {
     const id = parseInt(req.params.id);
 
     try {
+        const publicacion = await prisma.publicacion.findUnique({
+            where: { id }
+        });
+
+        if (!publicacion) {
+            return res.status(404).json({"error": "No se encontró una publicación con ese Id"});
+        }
+
         await prisma.publicacion.update({
             where: {
                 id: id
@@ -111,6 +143,14 @@ router.delete('/publicaciones/:id', async (req, res) => {
     const id = parseInt(req.params.id);
 
     try {
+        const publicacion = await prisma.publicacion.findUnique({
+            where: { id }
+        });
+
+        if (!publicacion) {
+            return res.status(404).json({"error": "No existe una publicación con ese Id"});
+        }
+
         await prisma.publicacion.delete({
             where: { id: id }
         });
